requirements.txt
streamlit
numpy
Pillow
tensorflow
scikit-learn
joblib
supabase
python-dotenv